+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.6.4 - Unreleased
 
+### Changed
+* `kwarray.embed_slice` now does error checking
+
 
 ## Version 0.6.3 - Released 2022-07-31